aider: Refactor spec file to use a mock `Ferrum::Browser` object for unit testing.
# Aider chat conversation:

USER: The `BankBuster` class builds on Vessel::Cargo, which is a crawler gem, which launches Chrome. What if we want to do unit tests/specs and don't want the crawler to launch Chrome? Or, if we could allow that, but we want to proxy requests that Chrome makes...

ASSISTANT: To prevent the crawler from launching Chrome during unit tests, you can use a technique called mocking or stubbing. This allows you to replace the actual `Ferrum` driver (which launches Chrome) with a mock object that simulates its behavior without actually launching a browser.

To do this, you would need to modify your test file (`spec/bank_buster_spec.rb`) to include a setup step that replaces the `Ferrum` driver with a mock object. You can use a gem like `rspec-mocks` to create the mock object.

For proxying requests, the `Ferrum` driver supports setting a proxy server. You can set the proxy server in your test setup, and then all requests made by the `Ferrum` driver will be routed through that proxy.

Here's a general idea of what the changes might look like:

```ruby
# spec/bank_buster_spec.rb
<<<<<<< ORIGINAL
require 'rspec'
require_relative '../bank_buster.rb'

describe BankBuster do
  let(:bank_buster) { BankBuster.new }
=======
require 'rspec'
require 'rspec/mocks'
require_relative '../bank_buster.rb'

describe BankBuster do
  let(:mock_driver) { instance_double("Ferrum::Browser") }
  let(:bank_buster) { BankBuster.new(driver: mock_driver) }
>>>>>>> UPDATED
```

In the updated code, we create a mock `Ferrum::Browser` object and pass it to the `BankBuster` constructor. This means that when the `BankBuster` class tries to use the `Ferrum` driver, it will actually be using our mock object, which won't launch a browser.

Please note that this is a simplified example and might not work as-is. You would need to adjust it to fit the specifics of your code and what you want to test.