aider: Add require statements for gems used in the `ElectricityStatsHandler` class.
# Aider chat conversation:

USER: Add the require statements for the gems used in the handler class. Check the `json_server.rb` require statements for 'inspiration'.

ASSISTANT: To add the require statements for the gems used in the `ElectricityStatsHandler` class, we need to modify the `handlers/electricity_stats_handler.rb` file.

Here is the change:

```ruby
handlers/electricity_stats_handler.rb
<<<<<<< ORIGINAL
# I maj 2023 använde vi 900 kWh och betalade 616 kr till elhandelsbolaget.
# Samt till elnätsbolaget 1299 kr - varav 467 kr är en statisk månadsavgift.
=======
require 'oj'
require 'awesome_print'
require 'pry'
require 'pry-nav'

# I maj 2023 använde vi 900 kWh och betalade 616 kr till elhandelsbolaget.
# Samt till elnätsbolaget 1299 kr - varav 467 kr är en statisk månadsavgift.
>>>>>>> UPDATED
```

This change adds the require statements for the 'oj', 'awesome_print', 'pry', and 'pry-nav' gems at the top of the file. These gems are used in the `ElectricityStatsHandler` class.