aider: Add the require statements for the 'faraday' and 'oj' gems in the `StravaWorkoutsHandler` class.
# Aider chat conversation:

USER: Add the require statements for the gems used in the handler class. Check the `json_server.rb` require statements for 'inspiration'.

ASSISTANT: To add the require statements for the gems used in the `StravaWorkoutsHandler` class, we need to modify the `handlers/strava_workouts_handler.rb` file.

Here is the change:

```ruby
handlers/strava_workouts_handler.rb
<<<<<<< ORIGINAL
class StravaWorkoutsHandler
  STRAVA_API_URL = 'https://www.strava.com/api/v3'
  CLIENT_ID = ENV['STRAVA_CLIENT_ID']
  CLIENT_SECRET = ENV['STRAVA_CLIENT_SECRET']
=======
require 'faraday'
require 'oj'

class StravaWorkoutsHandler
  STRAVA_API_URL = 'https://www.strava.com/api/v3'
  CLIENT_ID = ENV['STRAVA_CLIENT_ID']
  CLIENT_SECRET = ENV['STRAVA_CLIENT_SECRET']
>>>>>>> UPDATED
```

This change adds the require statements for the 'faraday' and 'oj' gems at the top of the file. These gems are used in the `StravaWorkoutsHandler` class.