aider: Added a `view` data property to `App.vue` to determine which view to show, and updated the template to conditionally render the button, QR code, and results table based on the value of `view`.
# Aider chat conversation:

USER: Just so we're clear, this is all part of one flow, so when you come to the home page, all you see is the button first, then all you see is the QR code of roughly the same size, then if it succeeds in fetching the data, it should only show the results table.

ASSISTANT: Understood. To implement this, we can add a `view` data property to our Vue app that will determine which view to show. This property can have three possible values: `'button'`, `'qrCode'`, and `'results'`.

Initially, `view` will be `'button'`, so only the `BigButton` component will be shown. When the button is clicked, `view` will be set to `'qrCode'`, so only the QR code will be shown. When the data is successfully fetched, `view` will be set to `'results'`, so only the results table will be shown.

Here's how you can modify the `App.vue` component to implement this:

```vue
frontend/App.vue
<<<<<<< ORIGINAL
<template>
  <div id="app">
    <BigButton v-if="!error" @click="handleClick" />
    <div v-if="error" class="error-notification">
      Något gick fel.
    </div>
    <!-- The results table will go here -->
  </div>
</template>

<script>
import BigButton from './components/BigButton.vue'

export default {
  name: 'App',
  components: {
    BigButton,
  },
  data() {
    return {
      error: null,
    }
  },
  methods: {
    handleClick() {
      // Handle the button click
      // If an error occurs, set this.error to the error message
    },
  },
}
</script>

<style scoped>
.error-notification {
  /* Add styles for the error notification */
}
</style>
=======
<template>
  <div id="app">
    <BigButton v-if="view === 'button'" @click="handleClick" />
    <div v-if="view === 'qrCode'">
      <!-- The QR code will go here -->
    </div>
    <div v-if="view === 'results'">
      <!-- The results table will go here -->
    </div>
    <div v-if="error" class="error-notification">
      Något gick fel.
    </div>
  </div>
</template>

<script>
import BigButton from './components/BigButton.vue'

export default {
  name: 'App',
  components: {
    BigButton,
  },
  data() {
    return {
      view: 'button',
      error: null,
    }
  },
  methods: {
    handleClick() {
      // Handle the button click
      this.view = 'qrCode';
      // If an error occurs, set this.error to the error message and this.view back to 'button'
    },
  },
}
</script>

<style scoped>
.error-notification {
  /* Add styles for the error notification */
}
</style>
>>>>>>> UPDATED
```

In this updated `App.vue` component, the `BigButton` component is only shown if `view` is `'button'`. The QR code is only shown if `view` is `'qrCode'`. The results table is only shown if `view` is `'results'`.

The `handleClick` method is updated to set `view` to `'qrCode'` when the button is clicked. If an error occurs during this process, you can set `this.error` to the error message and `this.view` back to `'button'` to show the error notification and go back to the button view.

Let me know if you need further assistance!