aider: Remove 'START' command from the `created` lifecycle hook in `frontend/src/App.vue`.
# Aider chat conversation:

USER: The START command, what does it do? Isn't it sent from multiple places? It should be sent only when clicking the big button, because as I remember it, it triggers the crawler to start the login process, no?

ASSISTANT: Yes, you're correct. The 'START' command is sent to the backend to initiate the login process. In your current setup, it's being sent in two places in `frontend/src/App.vue`:

1. When the WebSocket connection is opened in the `created` lifecycle hook.
2. When the big button is clicked in the `handleClick` method.

This could potentially cause the login process to be initiated twice. If you want the login process to start only when the big button is clicked, you should remove the 'START' command from the `created` lifecycle hook.

Here's the change needed in `frontend/src/App.vue`:

```vue
frontend/src/App.vue
<<<<<<< ORIGINAL
  created() {
    this.socket = this.$socket;
    this.socket.onmessage = this.handleMessage;
    this.socket.onopen = () => {
      this.socket.send('START')
    }
  },
=======
  created() {
    this.socket = this.$socket;
    this.socket.onmessage = this.handleMessage;
  },
>>>>>>> UPDATED
```

Please confirm this change before we proceed with the rest.